+++ src/HotelPosSystem.Tests/BookingListTests.cs
@@ -89,6 +89,34 @@ namespace HotelPosSystem.Tests {
             Assert.Null(commentLabel);
         }
 
+        [Fact]
+        public void BookingShouldBeRemoved() {
+            AutomationElement bookingList = Utilities.GetBookingListElement(_programWithDatabase, _automation);
+            Button removeButton = Utilities.GetElement(_automation, "removeButton2", bookingList).AsButton();
+
+            removeButton.Click();
+            Window modalWindow = _programWithDatabase.Application.GetAllTopLevelWindows(_automation)[0];
+            Button yesButton = modalWindow.FindFirstDescendant(descendant => descendant.ByName("Yes")).AsButton();
+            yesButton.Click();
+
+            Label? customerName = Utilities.GetElement(_automation, "customerName2", bookingList).AsLabel();
+            Assert.Null(customerName);
+        }
+
+        [Fact]
+        public void BookingShouldNotBeRemoved() {
+            AutomationElement bookingList = Utilities.GetBookingListElement(_programWithDatabase, _automation);
+            Button removeButton = Utilities.GetElement(_automation, "removeButton2", bookingList).AsButton();
+
+            removeButton.Click();
+            Window modalWindow = _programWithDatabase.Application.GetAllTopLevelWindows(_automation)[0];
+            Button noButton = modalWindow.FindFirstDescendant(descendant => descendant.ByName("No")).AsButton();
+            noButton.Click();
+
+            Label? customerName = Utilities.GetElement(_automation, "customerName2", bookingList).AsLabel();
+            Assert.NotNull(customerName);
+        }
+
         public void Dispose() {
             _automation.Dispose();
             _programWithDatabase.Dispose();