+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Layout improvements
+
 ## 0.3.0 - 2024-01-18
 
 ### Added