+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 
 ### Changed
 
-- Rename label "Paid for" to "Has paid"
+- Label for checkbox "Paid for" in booking form renamed to "Has paid"
 - Layout improvements
 - No customer/room is selected by default in booking form